Returning exrtma3d_fetchhpss.sh to original form.

+++ scripts/exrtma3d_fetchhpss.sh
@@ -19,36 +19,44 @@ postmsg "$jlogfile" "$msg"
 #========================= begin changes =======================================================
 envir=${envir}
 CYCLE=${PDY}${cyc}
+CYCLE_LAST=${CYCLE}
+CYCLE_p1=${CYCLE}
 mynoscrub=${COMIN}
 mynoscrubm1=${COMINm1}
+CYCLE_3=`$NDATE -0 $CYCLE`
+CYCLE_LAST=`$NDATE +1 $CYCLE`
 YYYYMMDDm1=`$NDATE -24 ${CYCLE} | cut -c 1-8`
 #=========================  end changes  =======================================================
 
-    i=1
-    YYYY=`echo $CYCLE | cut -c 1-4`
-    YYYYMM=`echo $CYCLE | cut -c 1-6`
-    YYYYMMDD=`echo $CYCLE | cut -c 1-8`
-    YYYYMMDDHH=`echo $CYCLE | cut -c 1-10`
-    MM=`echo $CYCLE | cut -c 5-6`
-    DD=`echo $CYCLE | cut -c 7-8`
-    HH=`echo $CYCLE | cut -c 9-10`
+i=0
+while [ $CYCLE_3 -le $CYCLE_LAST ] ; do
+
+    let "i=i+1"
+
+    YYYY=`echo $CYCLE_3 | cut -c 1-4`
+    YYYYMM=`echo $CYCLE_3 | cut -c 1-6`
+    YYYYMMDD=`echo $CYCLE_3 | cut -c 1-8`
+    YYYYMMDDHH=`echo $CYCLE_3 | cut -c 1-10`
+    MM=`echo $CYCLE_3 | cut -c 5-6`
+    DD=`echo $CYCLE_3 | cut -c 7-8`
+    HH=`echo $CYCLE_3 | cut -c 9-10`
     MN="00"
     MNp1="01"
     MNp2="02"
 
-    YY2=`echo $CYCLE | cut -c 3-4`
+    YY2=`echo $CYCLE_3 | cut -c 3-4`
     # JJJ=`/bin/date --date="${MM}/${DD}/${YYYY}" +"%j" `
     JJJ=`/bin/date --date="${YYYYMMDD}" +"%j" `
 
-    CYCLE_m1hr=`$NDATE -1 $CYCLE`
-    YYYY_m1hr=`echo $CYCLE_m1hr | cut -c 1-4`
-    YYYYMM_m1hr=`echo $CYCLE_m1hr | cut -c 1-6`
-    YYYYMMDD_m1hr=`echo $CYCLE_m1hr | cut -c 1-8`
-    YYYYMMDDHH_m1hr=`echo $CYCLE_m1hr | cut -c 1-10`
-    MM_m1hr=`echo $CYCLE_m1hr | cut -c 5-6`
-    DD_m1hr=`echo $CYCLE_m1hr | cut -c 7-8`
-    HH_m1hr=`echo $CYCLE_m1hr | cut -c 9-10`
-    YY2_m1hr=`echo $CYCLE_m1hr | cut -c 3-4`
+    CYCLE_3_m1hr=`$NDATE -1 $CYCLE_3`
+    YYYY_m1hr=`echo $CYCLE_3_m1hr | cut -c 1-4`
+    YYYYMM_m1hr=`echo $CYCLE_3_m1hr | cut -c 1-6`
+    YYYYMMDD_m1hr=`echo $CYCLE_3_m1hr | cut -c 1-8`
+    YYYYMMDDHH_m1hr=`echo $CYCLE_3_m1hr | cut -c 1-10`
+    MM_m1hr=`echo $CYCLE_3_m1hr | cut -c 5-6`
+    DD_m1hr=`echo $CYCLE_3_m1hr | cut -c 7-8`
+    HH_m1hr=`echo $CYCLE_3_m1hr | cut -c 9-10`
+    YY2_m1hr=`echo $CYCLE_3_m1hr | cut -c 3-4`
     # JJJ_m1hr=`/bin/date --date="${MM_m1hr}/${DD_m1hr}/${YYYY_m1hr}" +"%j" `
     JJJ_m1hr=`/bin/date --date="${YYYYMMDD_m1hr}" +"%j" `
 
@@ -145,17 +153,23 @@ YYYYMMDDm1=`$NDATE -24 ${CYCLE} | cut -c 1-8`
 #   if [ ! -d $wrkdir_nam ]; then
 #    mkdir -p  $wrkdir_nam
 #   fi
-    wrkdir_rap=${mynoscrub}/rap.${YYYYMMDD}     
-    if [ ! -d  $wrkdir_rap ]; then
+    if [ "$i" -eq 1 ] ; then 
+      wrkdir_rap=${mynoscrub}/rap.${YYYYMMDD}     
+      if [ ! -d  $wrkdir_rap ]; then
        mkdir -p  $wrkdir_rap
+      fi
     fi
-    wrkdir_radar=${mynoscrub}/radar.${YYYYMMDD}     
-    if [ ! -d  $wrkdir_radar ]; then
+    if [ "$i" -eq 1 ] ; then 
+      wrkdir_radar=${mynoscrub}/radar.${YYYYMMDD}     
+      if [ ! -d  $wrkdir_radar ]; then
        mkdir -p  $wrkdir_radar
+      fi
     fi
-    wrkdir_lightning=${mynoscrub}/lightning.${YYYYMMDD}     
-    if [ ! -d  $wrkdir_lightning ]; then
+    if [ "$i" -eq 1 ] ; then 
+      wrkdir_lightning=${mynoscrub}/lightning.${YYYYMMDD}     
+      if [ ! -d  $wrkdir_lightning ]; then
        mkdir -p  $wrkdir_lightning
+      fi
     fi
 #   wrkdir_rtma=${mynoscrub}/${NET}.${YYYYMMDD}   
 #   if [ ! -d  $wrkdir_rtma ]; then
@@ -185,28 +199,30 @@ YYYYMMDDm1=`$NDATE -24 ${CYCLE} | cut -c 1-8`
 #==================================================================================================
 
 #==================================================================================================
-    cd $wrkdir_rap
+    if [ "$i" -eq 1 ] ; then
+      cd $wrkdir_rap
 
-    /bin/rm -rf select_list_1.txt_tmp
-    /bin/rm -rf select_list_1.txt
+      /bin/rm -rf select_list_1.txt_tmp
+      /bin/rm -rf select_list_1.txt
 
-    htar -tvf ${hpsspath}/${tarfile_rap_bufr}  > list_all_1.txt
+      htar -tvf ${hpsspath}/${tarfile_rap_bufr}  > list_all_1.txt
 
-    cat list_all_1.txt | grep rap.t${HH}z.prepbufr.tm00 >> select_list_1.txt_tmp
-    cat list_all_1.txt | grep rap.t${HH}z.lgycld.tm00.bufr_d >> select_list_1.txt_tmp
-    cat list_all_1.txt | grep rap.t${HH}z.lghtng.tm00.bufr_d >> select_list_1.txt_tmp
+      cat list_all_1.txt | grep rap.t${HH}z.prepbufr.tm00 >> select_list_1.txt_tmp
+      cat list_all_1.txt | grep rap.t${HH}z.lgycld.tm00.bufr_d >> select_list_1.txt_tmp
+      cat list_all_1.txt | grep rap.t${HH}z.lghtng.tm00.bufr_d >> select_list_1.txt_tmp
 
-    nlines=`wc -l select_list_1.txt_tmp`
-    nlines=${nlines% select*}
-    echo "nlines ="$nlines
+      nlines=`wc -l select_list_1.txt_tmp`
+      nlines=${nlines% select*}
+      echo "nlines ="$nlines
 
-    it=1
-    while [ $it -le $nlines ] ; do
-      var="`cat select_list_1.txt_tmp | head -n $it | tail  -1`" 
-      echo "./"${var#* ./} >> select_list_1.txt
-      let "it=it+1"
-    done
-    htar -xvf ${hpsspath}/${tarfile_rap_bufr} -L select_list_1.txt
+      it=1
+      while [ $it -le $nlines ] ; do
+        var="`cat select_list_1.txt_tmp | head -n $it | tail  -1`" 
+        echo "./"${var#* ./} >> select_list_1.txt
+        let "it=it+1"
+      done
+      htar -xvf ${hpsspath}/${tarfile_rap_bufr} -L select_list_1.txt
+    fi
         
 #==================================================================================================
     cd $wrkdir_hrrr
@@ -216,11 +232,15 @@ YYYYMMDDm1=`$NDATE -24 ${CYCLE} | cut -c 1-8`
 
     htar -tvf ${hpsspath_hrrr}/${tarfile_hrrr_init}  > list_all_1.txt
 
-    cat list_all_1.txt | grep hrrr.t${HH}z.NSSLRefInGSI.bufr >> select_list_1.txt_tmp
-    cat list_all_1.txt | grep hrrr.t${HH}z.NASALaRCCloudInGSI.bufr >> select_list_1.txt_tmp
+    if [ "$i" -eq 1 ] ; then
+      cat list_all_1.txt | grep hrrr.t${HH}z.NSSLRefInGSI.bufr >> select_list_1.txt_tmp
+      cat list_all_1.txt | grep hrrr.t${HH}z.NASALaRCCloudInGSI.bufr >> select_list_1.txt_tmp
 #     cat list_all_1.txt | grep "\bhrrr.t${HH}z.wrfguess\b"  >> select_list_1.txt_tmp
 #     cat list_all_1.txt | grep "\<hrrr.t${HH}z.wrfguess\>"  >> select_list_1.txt_tmp
-    cat list_all_1.txt | grep -w hrrr.t${HH}z.wrfguess  >> select_list_1.txt_tmp
+      cat list_all_1.txt | grep -w hrrr.t${HH}z.wrfguess  >> select_list_1.txt_tmp
+    else
+      cat list_all_1.txt | grep -w hrrr.t${HH}z.wrfguess_rap >> select_list_1.txt_tmp
+    fi
 
     nlines=`wc -l select_list_1.txt_tmp`
     nlines=${nlines% select*}
@@ -235,35 +255,36 @@ YYYYMMDDm1=`$NDATE -24 ${CYCLE} | cut -c 1-8`
     htar -xvf ${hpsspath_hrrr}/${tarfile_hrrr_init} -L select_list_1.txt
         
 #==================================================================================================
-    cd $wrkdir_radar
-    rm -rf $mrmsradar_grib2_dir
-    if [ ! -d $mrmsradar_grib2_dir ] ; then
-      mkdir -p $mrmsradar_grib2_dir
-    fi
+    if [ "$i" -eq 1 ] ; then
+      cd $wrkdir_radar
+      rm -rf $mrmsradar_grib2_dir
+      if [ ! -d $mrmsradar_grib2_dir ] ; then
+        mkdir -p $mrmsradar_grib2_dir
+      fi
 
-    hsi get "${mrmsradar_dir}/${zipfile_mrms_3hr}"
+      hsi get "${mrmsradar_dir}/${zipfile_mrms_3hr}"
 
-    /bin/rm -rf select_list_1.txt_tmp
-    /bin/rm -rf select_list_1.txt
-    /bin/rm -rf list_all_1.txt
-    /bin/rm -rf select_list_1.txt_total
+      /bin/rm -rf select_list_1.txt_tmp
+      /bin/rm -rf select_list_1.txt
+      /bin/rm -rf list_all_1.txt
+      /bin/rm -rf select_list_1.txt_total
 
-    /bin/unzip -v ${zipfile_mrms_3hr} > list_all_1.txt
+      /bin/unzip -v ${zipfile_mrms_3hr} > list_all_1.txt
 
-    zfiles="${zipfile_mrms_1mi} ${zipfile_mrms_1mip1} ${zipfile_mrms_1mip2}"
-    for zf in $zfiles
-    do
-      nlines=`cat list_all_1.txt | awk {'print $8'} | grep "${zf}" | wc -l`
-      if [ $nlines -eq 1 ]
-      then
-        /bin/unzip ${zipfile_mrms_3hr} "${zf}" -d ${mrmsradar_grib2_dir}
+      zfiles="${zipfile_mrms_1mi} ${zipfile_mrms_1mip1} ${zipfile_mrms_1mip2}"
+      for zf in $zfiles
+      do
+        nlines=`cat list_all_1.txt | awk {'print $8'} | grep "${zf}" | wc -l`
+        if [ $nlines -eq 1 ]
+        then
+          /bin/unzip ${zipfile_mrms_3hr} "${zf}" -d ${mrmsradar_grib2_dir}
 
-        /bin/rm -rf select_list_1.txt_tmp
-        /bin/rm -rf select_list_1.txt
-        /bin/unzip -v ${mrmsradar_grib2_dir}/${zf} > select_list_1.txt_tmp
-         cat select_list_1.txt_tmp | grep "MRMS_MergedReflectivityQC_" | awk {'if($8 !~ /conusPlus/) print $8'} > select_list_1.txt
-         nl=`cat select_list_1.txt | wc -l `
-         if [ $nl -ge 1 ]
+          /bin/rm -rf select_list_1.txt_tmp
+          /bin/rm -rf select_list_1.txt
+          /bin/unzip -v ${mrmsradar_grib2_dir}/${zf} > select_list_1.txt_tmp
+          cat select_list_1.txt_tmp | grep "MRMS_MergedReflectivityQC_" | awk {'if($8 !~ /conusPlus/) print $8'} > select_list_1.txt
+          nl=`cat select_list_1.txt | wc -l `
+          if [ $nl -ge 1 ]
           then
             cat select_list_1.txt | xargs -d '\n' unzip ${mrmsradar_grib2_dir}/${zf} -d ${mrmsradar_grib2_dir}
           fi
@@ -271,64 +292,74 @@ YYYYMMDDm1=`$NDATE -24 ${CYCLE} | cut -c 1-8`
           cat select_list_1.txt >> select_list_1.txt_total
           echo "=================================="  >> select_list_1.txt_total
         fi
-    done
+      done
 
       /bin/rm -f ${zipfile_mrms_3hr}
       
+    fi
         
 #==================================================================================================
 
-    cd $wrkdir_lightning
+    if [ "$i" -eq 1 ] ; then
+      cd $wrkdir_lightning
 
-    for lghtn_dir in $lghtn_entln_dir $lghtn_vaisala_dir
-    do
-      rm -rf $lghtn_dir
-      if [ ! -d $lghtn_dir ] ; then
-        mkdir -p $lghtn_dir
-      fi
-
-      if [ $lghtn_dir = $lghtn_entln_dir ] ; then
-        hsi get "${entln_dir}/${zipfile_lghtn_1dd}"
-      elif [ $lghtn_dir = $lghtn_vaisala_dir ] ; then
-        hsi get "${vaisala_dir}/${zipfile_lghtn_1dd}"
-      fi
-
-      /bin/rm -rf select_list_1.txt_tmp
-      /bin/rm -rf select_list_1.txt
-      /bin/rm -rf list_all_1.txt
-      /bin/rm -rf select_list_1.txt_total
-
-      /bin/unzip -v ${zipfile_lghtn_1dd} > list_all_1.txt
-
-      if [ "${fnmhead_lghtn}" = "${fnmhead_lghtn_m1hr}" ] ; then
-        fheads="${fnmhead_lghtn}"
-      else
-        fheads="${fnmhead_lghtn} ${fnmhead_lghtn_m1hr} "
-      fi
-      for fh in $fheads
+      for lghtn_dir in $lghtn_entln_dir $lghtn_vaisala_dir
       do
-        cat list_all_1.txt | grep "${fh}"  > select_list_1.txt_tmp
-        cat select_list_1.txt_tmp | awk '{print $8}' > select_list_1.txt
-        nlines=`cat select_list_1.txt | wc -l `
-        if [ $nlines -ge 1 ]
-        then
-          cat select_list_1.txt | xargs -d '\n' unzip ${zipfile_lghtn_1dd} -d ${lghtn_dir}
+        rm -rf $lghtn_dir
+        if [ ! -d $lghtn_dir ] ; then
+          mkdir -p $lghtn_dir
         fi
-        cat select_list_1.txt >> select_list_1.txt_total
-        echo "=================================="  >> select_list_1.txt_total
+
+        if [ $lghtn_dir = $lghtn_entln_dir ] ; then
+          hsi get "${entln_dir}/${zipfile_lghtn_1dd}"
+        elif [ $lghtn_dir = $lghtn_vaisala_dir ] ; then
+          hsi get "${vaisala_dir}/${zipfile_lghtn_1dd}"
+        fi
+
+        /bin/rm -rf select_list_1.txt_tmp
+        /bin/rm -rf select_list_1.txt
+        /bin/rm -rf list_all_1.txt
+        /bin/rm -rf select_list_1.txt_total
+
+        /bin/unzip -v ${zipfile_lghtn_1dd} > list_all_1.txt
+
+        if [ "${fnmhead_lghtn}" = "${fnmhead_lghtn_m1hr}" ] ; then
+          fheads="${fnmhead_lghtn}"
+        else
+          fheads="${fnmhead_lghtn} ${fnmhead_lghtn_m1hr} "
+        fi
+        for fh in $fheads
+        do
+          cat list_all_1.txt | grep "${fh}"  > select_list_1.txt_tmp
+          cat select_list_1.txt_tmp | awk '{print $8}' > select_list_1.txt
+          nlines=`cat select_list_1.txt | wc -l `
+          if [ $nlines -ge 1 ]
+          then
+            cat select_list_1.txt | xargs -d '\n' unzip ${zipfile_lghtn_1dd} -d ${lghtn_dir}
+          fi
+          cat select_list_1.txt >> select_list_1.txt_total
+          echo "=================================="  >> select_list_1.txt_total
+        done
+
+        /bin/rm -f ${zipfile_lghtn_1dd}
+
       done
-
-      /bin/rm -f ${zipfile_lghtn_1dd}
-
-    done
       
+    fi
         
+#==================================================================================================
+
+#   cd $tmp_1
+#==================================================================================================
+
 
 #==================================================================================================
-#==> rm tmp_1 directory
+#==> rm tmp_1 directory + advance CYCLE
 #==================================================================================================
     cd ${mynoscrub}
 
+    CYCLE_3=`$NDATE +1 $CYCLE_3`
+done
 
 
 ################################################################################################
